[refactor] Compose 관련 패키지 분리 작업
