+++ app/src/main/java/com/mashup/zuzu/compose/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.mashup.zuzu.ui.theme
+package com.mashup.zuzu.compose.theme
 
 import androidx.compose.runtime.*
 import androidx.compose.ui.graphics.Color
