+++ app/src/main/java/com/mashup/zuzu/compose/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.mashup.zuzu.ui.theme
+package com.mashup.zuzu.compose.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
