+++ app/src/main/java/com/mashup/zuzu/ui/category/Category.kt
@@ -28,10 +28,10 @@ import com.mashup.zuzu.data.model.Category
 import com.mashup.zuzu.data.model.Wine
 import com.mashup.zuzu.data.model.categoryList
 import com.mashup.zuzu.data.model.wines
-import com.mashup.zuzu.ui.component.*
-import com.mashup.zuzu.ui.component.TabPosition
-import com.mashup.zuzu.ui.theme.Black
-import com.mashup.zuzu.ui.theme.ProofTheme
+import com.mashup.zuzu.compose.component.*
+import com.mashup.zuzu.compose.component.TabPosition
+import com.mashup.zuzu.compose.theme.Black
+import com.mashup.zuzu.compose.theme.ProofTheme
 
 /**
  * @Created by 김현국 2022/07/03
