+++ app/src/main/java/com/mashup/zuzu/ui/home/Home.kt
@@ -29,8 +29,8 @@ import com.mashup.zuzu.R
 import com.mashup.zuzu.data.model.BestWorldCup
 import com.mashup.zuzu.data.model.Wine
 import com.mashup.zuzu.data.model.categoryList
-import com.mashup.zuzu.ui.component.*
-import com.mashup.zuzu.ui.theme.ProofTheme
+import com.mashup.zuzu.compose.component.*
+import com.mashup.zuzu.compose.theme.ProofTheme
 import timber.log.Timber
 
 /**
