+++ app/src/main/java/com/mashup/zuzu/ui/setting/Setting.kt
@@ -19,7 +19,7 @@ import coil.compose.AsyncImage
 import com.mashup.zuzu.R
 import com.mashup.zuzu.data.model.User
 import com.mashup.zuzu.data.model.user
-import com.mashup.zuzu.ui.theme.ProofTheme
+import com.mashup.zuzu.compose.theme.ProofTheme
 import com.mashup.zuzu.ui.user.UserUiState
 
 /**
